// model/audioOutputFileURLResponse.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class AudioOutputFileURLResponse {
    /**
    * A presigned URL that points to the output audio file.
    */
    'outputUrl'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "outputUrl",
            "baseName": "output_url",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AudioOutputFileURLResponse.attributeTypeMap;
    }
}