// model/audioSeparationRunInfoResponse.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class AudioSeparationRunInfoResponse {
    /**
    * The URL pointing to the generated audio file for the detected foreground audio.
    */
    'foregroundAudioUrl'?: string;
    /**
    * The URL pointing to the generated audio file for the detected background audio.
    */
    'backgroundAudioUrl'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "foregroundAudioUrl",
            "baseName": "foreground_audio_url",
            "type": "string"
        },
        {
            "name": "backgroundAudioUrl",
            "baseName": "background_audio_url",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AudioSeparationRunInfoResponse.attributeTypeMap;
    }
}