// model/bodyTranslateTranslatePost.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { Formalities } from './formalities';
import { Gender } from './gender';
import { Languages } from './languages';

export class BodyTranslateTranslatePost {
    'sourceLanguage': Languages;
    'targetLanguage': Languages;
    'texts': Array<string>;
    'age'?: number;
    /**
    * Specifies whether the translated text should be formal or informal. For example, `1` (formal) may use polite language, while `2` (informal) may use casual or colloquial language.
    */
    'formality'?: Formalities;
    'gender'?: Gender;
    /**
    * An optional list of dictionary IDs selected by the user. Each entry must be an integer corresponding to a valid dictionary ID. If provided, at least one ID is required.
    */
    'chosenDictionaries'?: Array<number>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "sourceLanguage",
            "baseName": "source_language",
            "type": "Languages"
        },
        {
            "name": "targetLanguage",
            "baseName": "target_language",
            "type": "Languages"
        },
        {
            "name": "texts",
            "baseName": "texts",
            "type": "Array<string>"
        },
        {
            "name": "age",
            "baseName": "age",
            "type": "number"
        },
        {
            "name": "formality",
            "baseName": "formality",
            "type": "Formalities"
        },
        {
            "name": "gender",
            "baseName": "gender",
            "type": "Gender"
        },
        {
            "name": "chosenDictionaries",
            "baseName": "chosen_dictionaries",
            "type": "Array<number>"
        }    ];

    static getAttributeTypeMap() {
        return BodyTranslateTranslatePost.attributeTypeMap;
    }
}

export namespace BodyTranslateTranslatePost {
}