// model/createAPIKeyRequestPayload.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CreateAPIKeyRequestPayload {
    'name': string;
    'teamId': number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "teamId",
            "baseName": "team_id",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return CreateAPIKeyRequestPayload.attributeTypeMap;
    }
}