// model/createCustomVoiceOut.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CreateCustomVoiceOut {
    'voiceId'?: number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "voiceId",
            "baseName": "voice_id",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return CreateCustomVoiceOut.attributeTypeMap;
    }
}