// model/createTTSRequestPayload.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { Gender } from './gender';
import { Languages } from './languages';

export class CreateTTSRequestPayload {
    /**
    * The text to be converted to speech
    */
    'text': string;
    /**
    * The voice ID to be used to generate speech.
    */
    'voiceId': number;
    /**
    * The language in which the provided text is written in.
    */
    'language': Languages;
    'gender'?: Gender;
    'age'?: any | null;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "text",
            "baseName": "text",
            "type": "string"
        },
        {
            "name": "voiceId",
            "baseName": "voice_id",
            "type": "number"
        },
        {
            "name": "language",
            "baseName": "language",
            "type": "Languages"
        },
        {
            "name": "gender",
            "baseName": "gender",
            "type": "Gender"
        },
        {
            "name": "age",
            "baseName": "age",
            "type": "any"
        }    ];

    static getAttributeTypeMap() {
        return CreateTTSRequestPayload.attributeTypeMap;
    }
}

export namespace CreateTTSRequestPayload {
}