// model/createTTSStreamRequestPayload.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { Gender } from './gender';
import { Languages } from './languages';
import { TTSStreamOutputFormat } from './tTSStreamOutputFormat';

export class CreateTTSStreamRequestPayload {
    /**
    * The content you want converted into spoken audio. This can be anything from a single sentence to paragraphs of text, supporting punctuation for natural speech patterns.
    */
    'text': string;
    /**
    * The unique identifier for your selected voice profile. You can obtain available voice IDs from the `/list-voices` endpoint or create custom voices with the `/create-custom-voice` endpoint.
    */
    'voiceId': number;
    /**
    * The source language of your input text. This helps the system apply the correct pronunciation rules and speech patterns.
    */
    'language': Languages;
    /**
    * The preferred gender characteristics of the synthesized voice (`0 = Not Specified, 1 = Male, 2 = Female, 9 = Not Applicable`). Defaults to `null`.
    */
    'gender'?: Gender;
    /**
    * The approximate age (between 1-100 years) to be reflected in the voice characteristics. This parameter helps fine-tune the timbre and speech patterns to match different age groups.
    */
    'age'?: number;
    /**
    * The audio file format for the generated speech stream.
    */
    'outputFormat'?: TTSStreamOutputFormat;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "text",
            "baseName": "text",
            "type": "string"
        },
        {
            "name": "voiceId",
            "baseName": "voice_id",
            "type": "number"
        },
        {
            "name": "language",
            "baseName": "language",
            "type": "Languages"
        },
        {
            "name": "gender",
            "baseName": "gender",
            "type": "Gender"
        },
        {
            "name": "age",
            "baseName": "age",
            "type": "number"
        },
        {
            "name": "outputFormat",
            "baseName": "output_format",
            "type": "TTSStreamOutputFormat"
        }    ];

    static getAttributeTypeMap() {
        return CreateTTSStreamRequestPayload.attributeTypeMap;
    }
}

export namespace CreateTTSStreamRequestPayload {
}