// model/createTextToVoiceRequestPayload.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CreateTextToVoiceRequestPayload {
    /**
    * The text content that will be converted into synthesized speech.
    */
    'text': string;
    /**
    * A detailed description of the desired voice characteristics to guide the synthesis engine. You can specify attributes such as gender, age, accent, emotional tone, speaking style, or cultural context to create authentic voices.
    */
    'voiceDescription': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "text",
            "baseName": "text",
            "type": "string"
        },
        {
            "name": "voiceDescription",
            "baseName": "voice_description",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return CreateTextToVoiceRequestPayload.attributeTypeMap;
    }
}