// model/createTranslatedStoryRequestPayload.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { Languages } from './languages';

export class CreateTranslatedStoryRequestPayload {
    /**
    * The lang
    */
    'targetLanguage'?: Languages;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "targetLanguage",
            "baseName": "target_language",
            "type": "Languages"
        }    ];

    static getAttributeTypeMap() {
        return CreateTranslatedStoryRequestPayload.attributeTypeMap;
    }
}

export namespace CreateTranslatedStoryRequestPayload {
}