// model/createTranslationStreamRequestPayload.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { Formalities } from './formalities';
import { Gender } from './gender';
import { Languages } from './languages';

export class CreateTranslationStreamRequestPayload {
    'text': string;
    'sourceLanguage': Languages;
    'targetLanguage': Languages;
    'formality'?: Formalities | null;
    'gender'?: Gender | null;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "text",
            "baseName": "text",
            "type": "string"
        },
        {
            "name": "sourceLanguage",
            "baseName": "source_language",
            "type": "Languages"
        },
        {
            "name": "targetLanguage",
            "baseName": "target_language",
            "type": "Languages"
        },
        {
            "name": "formality",
            "baseName": "formality",
            "type": "Formalities"
        },
        {
            "name": "gender",
            "baseName": "gender",
            "type": "Gender"
        }    ];

    static getAttributeTypeMap() {
        return CreateTranslationStreamRequestPayload.attributeTypeMap;
    }
}

export namespace CreateTranslationStreamRequestPayload {
}