// model/dialogueItem.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* A JSON that represents a single piece of dialogue or speech within a given time range. It includes details about the timing (start and end), the content of the dialogue (text), and the speaker who delivers the dialogue
*/
export class DialogueItem {
    /**
    * The start time (in seconds) of the dialogue.
    */
    'start': number;
    /**
    * The end time (in seconds) of the dialogue.
    */
    'end': number;
    /**
    * The actual text content of the dialogue spoken by the speaker.
    */
    'text': string;
    /**
    * The name or identifier of the speaker delivering the dialogue.
    */
    'speaker': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "start",
            "baseName": "start",
            "type": "number"
        },
        {
            "name": "end",
            "baseName": "end",
            "type": "number"
        },
        {
            "name": "text",
            "baseName": "text",
            "type": "string"
        },
        {
            "name": "speaker",
            "baseName": "speaker",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return DialogueItem.attributeTypeMap;
    }
}